sanitize: null writes, write errors, oversized int,
and set incoming_end to size_t

+++ src/connection.h
@@ -75,7 +75,7 @@ typedef struct connection {
 	time_t connect_time;
 	time_t timeout;
 	char *incoming;
-	unsigned incoming_end;
+	size_t incoming_end;
 	list_t *outgoing;
 	char *partial;
 	list_t *incoming_lines;
